👩‍💻Organize file stucture & Fixed readme.txt
