+++ src/components/readme.txt
@@ -1,0 +1,6 @@
+components
+
+for small react components that will later gather with another components
+ect.
+List.tsx
+Form.tsx
