+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
-import App from './App'
+import App from './routes/App'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import './assets/scss/_main.scss'
