+++ src/utils/readme.txt
@@ -1,4 +1,4 @@
-Utils
+Utils (functions and interfaces)
 
 I always create a lot of functions within a component.
 As a result, the component can become quite bloated.
